%
% Copyright (C) 2010, 23andMe, Inc.
% 
% Written by David A. Hinds <dhinds@sonic.net>
% 
% This is free software; you can redistribute it and/or modify it
% under the terms of the GNU General Public License as published by
% the Free Software Foundation; either version 3 of the license, or
% (at your option) any later version.
% 
% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU General Public License for more details.
% 
% You should have received a copy of the GNU General Public License
% along with this program.  If not, see <http://www.gnu.org/licenses/>
% 
\name{adjust.ld}
\alias{adjust.ld}
\title{Adjust Model to Account for Incomplete Linkage Disequilibrium}
\description{
  Adjust disease model parameters to account for incomplete linkage
  disequilibrium between the causal locus and an assayed marker.
}
\usage{
adjust.ld(model, m, dprime, rsqr)
}
\arguments{
  \item{model}{a model as generated by \code{\link{cc.model}}.}
  \item{m}{frequency of the marker allele in positive disequilibrium
    with the risk allele.}
  \item{dprime, rsqr}{alternate scales for expressing linkage
    disequilibrium between the causal locus and the marker.  One should
    be specified.}
}
\details{
  While \code{dprime} and \code{rsqr} can both vary from 0 to 1, the
  maximum possible value of \code{rsqr} is less than 1 if the
  frequencies of the causal and marker loci are unequal.  \code{dprime}
  is scaled so that a value of 1 represents the maximum disequilibrium
  given the specified frequencies.
}
\value{
  A new model as returned by \code{\link{cc.model}}, updated to account
  for incomplete linkage disequilibrium.  The original model is stored
  in a list element named \code{'orig.model'}.
}
\seealso{
  \code{\link{cc.model}}, \code{\link{cc.power}}.
}
\examples{
m <- cc.model(0.1, prevalence=0.2, odds.ratio=1.5)
adjust.ld(m, 0.2, dprime=0.8)
}
\keyword{manip}
