%
% Copyright (C) 2009, Perlegen Sciences, Inc.
%
% Written by David A. Hinds <dhinds@sonic.net>
%
% This is free software; you can redistribute it and/or modify it
% under the terms of the GNU General Public License as published by
% the Free Software Foundation; either version 3 of the license, or
% (at your option) any later version.
%
% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU General Public License for more details.
%
% You should have received a copy of the GNU General Public License
% along with this program.  If not, see <http://www.gnu.org/licenses/>
%
\name{demo_01}
\docType{data}
\alias{assay.dat.01}
\alias{assay.def.01}
\alias{assay.pos.01}
\alias{samples.01}
\title{Genotype Data for 5000 SNPs on 270 HapMap Samples}
\description{
  Genotype data for 5000 SNPs distributed uniformly across the genome,
  for 270 samples from the International HapMap Project.  This is a
  subset of data generated by Perlegen for the GAIN collaboration,
  using a set of SNP arrays designed to tag LD bins across the genome.
}
\usage{data(demo_01)}
\source{Perlegen Sciences, Inc., publicly released data.}
\references{
  The GAIN Collaborative Research Group (2007) New models of
  collaboration in genome-wide association studies: the Genetic
  Association Information Network.  \emph{Nat. Genet.} \bold{39}:
  1045-1051.

  The Database of Genotypes and Phenotypes.
  \url{http://www.ncbi.nlm.nih.gov/sites/entrez?db=gap}.

  The International HapMap Project. \url{http://www.hapmap.org}.
}
\examples{\dontrun{
# load demo datasets into GT.DB
demo('setup.gt.demo')
}}
\keyword{datasets}
