%
% Copyright (C) 2009, Perlegen Sciences, Inc.
%
% Written by David A. Hinds <dhinds@sonic.net>
%
% This is free software; you can redistribute it and/or modify it
% under the terms of the GNU General Public License as published by
% the Free Software Foundation; either version 3 of the license, or
% (at your option) any later version.
%
% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU General Public License for more details.
%
% You should have received a copy of the GNU General Public License
% along with this program.  If not, see <http://www.gnu.org/licenses/>
%
\name{demo_02}
\docType{data}
\alias{assay.dat.02}
\alias{assay.def.02}
\alias{assay.pos.02}
\alias{samples.02}
\title{Genotype Data for 371 Chr21 SNPs on 275 HapMap Samples}
\description{
  Genotype data for 371 SNPs distributed across a 300kb segment of human
  chromosome 21, for 275 samples from the International HapMap Project.
  The data was generated for the Phase II HapMap by Perlegen Sciences,
  using a custom array set developed for that project.
}
\usage{data(demo_02)}
\source{Perlegen Sciences, Inc., publicly released data.}
\references{
  The International HapMap Consortium (2007).  A second generation human
  haplotype map of over 3.1 million SNPs.  \emph{Nature} \bold{449}:
  851-861.

  The International HapMap Project. \url{http://www.hapmap.org}.
}
\keyword{datasets}
